Relocate contact details and update documentation link in README
The contact details have been moved from the top to the bottom of the README file. Additionally, the text for the documentation link has been modified to imply that it will be updated with more examples in the future. Lastly, a correction was made to the git clone URL guiding contributors to clone from the correct original repository.